+++ lib/core/base/app_base.dart
@@ -1,14 +1,8 @@
-import 'package:calvesia/feature/Authencitation/viewmodel/user_view_model.dart';
-import 'package:calvesia/feature/provider/base_provider.dart';
-import 'package:calvesia/feature/provider/explore_page_provider.dart';
-import 'package:calvesia/feature/provider/header_provider.dart';
-import 'package:calvesia/feature/provider/post_provider.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import '../../config/thema.dart';
 import '../../feature/onboard/on_board_page.dart';
 import '../../feature/pages/base_page.dart';
-import 'package:provider/provider.dart';
 
 class MyApp extends StatefulWidget {
   const MyApp({Key? key}) : super(key: key);
@@ -51,18 +45,11 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [
-        ChangeNotifierProvider(create: (_) => PostShareProvider()),
-        ChangeNotifierProvider(
-            create: (_) => PostIsSharingAndShowingProvider()),
-      ],
-      child: MaterialApp(
-        debugShowCheckedModeBanner: false,
-        title: 'Flutter Demo',
-        theme: baseThemeData(),
-        home: isSplasing ? splashScreen() : home(),
-      ),
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Flutter Demo',
+      theme: baseThemeData(),
+      home: isSplasing ? splashScreen() : home(),
     );
   }
 }
