+++ pubspec.yaml
@@ -42,7 +42,6 @@ dependencies:
   firebase_storage: ^10.2.15
   image_picker: ^0.8.5
   flutter_svg: ^1.0.3
-  provider: ^6.0.2
   intl:
   flutterfire_ui: ^0.4.1+2
   flutter_hooks: ^0.18.0
